soapsales/requirements.txt
asgiref==3.2.7
Babel==2.8.0
cffi==1.14.0
cryptography==2.9.2
Django==3.0.7
django-cors-headers==3.4.0
django-extra-fields==2.0.5
django-filter==2.3.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==2.1.2
django-prices==2.2.0
django-rest-knox==4.1.0
django-reversion==3.0.7
django-simple-history==2.10.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
drf-writable-nested==0.6.0
enmerkar==0.7.1
importlib-metadata==1.6.1
Markdown==3.2.2
Pillow==7.1.2
prices==1.0.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
satchless==1.2.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
typing==3.7.4.1
zipp==3.1.0
